+++ threedi_modelchecker/config.py
@@ -108,6 +108,12 @@ groundwater_settings_id = (
     .scalar_subquery()
 )
 groundwater_filter = models.GroundWater.id == groundwater_settings_id
+vegetation_drag_settings_id = (
+    Query(models.GlobalSetting.vegetation_drag_settings_id)
+    .filter(first_setting_filter)
+    .scalar_subquery()
+)
+vegetation_drag_filter = models.VegetationDrag.id == vegetation_drag_settings_id
 
 CONDITIONS = {
     "has_dem": Query(models.GlobalSetting).filter(
@@ -1206,6 +1212,7 @@ CHECKS += [
         (models.GroundWater, models.GlobalSetting.groundwater_settings_id),
         (models.NumericalSettings, models.GlobalSetting.numerical_settings_id),
         (models.ControlGroup, models.GlobalSetting.control_group_id),
+        (models.VegetationDrag, models.GlobalSetting.vegetation_drag_settings_id),
     )
 ]
 
@@ -1236,6 +1243,7 @@ CHECKS += [
             models.GlobalSetting.interflow_settings_id,
             models.GlobalSetting.simple_infiltration_settings_id,
             models.GlobalSetting.groundwater_settings_id,
+            models.GlobalSetting.vegetation_drag_settings_id,
         ]
     )
 ]
@@ -1644,6 +1652,121 @@ CHECKS += [
     ),
 ]
 
+## 05xx: VEGETATION DRAG
+CHECKS += [
+    RangeCheck(
+        error_code=501,
+        column=models.VegetationDrag.height,
+        filters=vegetation_drag_filter,
+        min_value=0,
+        left_inclusive=False,
+    ),
+    QueryCheck(
+        error_code=502,
+        column=models.VegetationDrag.height,
+        invalid=Query(models.VegetationDrag).filter(
+            vegetation_drag_filter,
+            models.VegetationDrag.height == None,
+            is_none_or_empty(models.VegetationDrag.height_file),
+        ),
+        message="v2_vegetation_drag.height must be defined.",
+    ),
+    QueryCheck(
+        error_code=503,
+        level=CheckLevel.WARNING,
+        column=models.VegetationDrag.height,
+        invalid=Query(models.VegetationDrag).filter(
+            vegetation_drag_filter,
+            models.VegetationDrag.height == None,
+            ~is_none_or_empty(models.VegetationDrag.height_file),
+        ),
+        message="v2_vegetation_drag.height is recommended as fallback value when using a height_file.",
+    ),
+    RangeCheck(
+        error_code=504,
+        column=models.VegetationDrag.stem_count,
+        filters=vegetation_drag_filter,
+        min_value=0,
+        left_inclusive=False,
+    ),
+    QueryCheck(
+        error_code=505,
+        column=models.VegetationDrag.stem_count,
+        invalid=Query(models.VegetationDrag).filter(
+            vegetation_drag_filter,
+            models.VegetationDrag.stem_count == None,
+            is_none_or_empty(models.VegetationDrag.stem_count_file),
+        ),
+        message="v2_vegetation_drag.stem_count must be defined.",
+    ),
+    QueryCheck(
+        error_code=506,
+        level=CheckLevel.WARNING,
+        column=models.VegetationDrag.stem_count,
+        invalid=Query(models.VegetationDrag).filter(
+            vegetation_drag_filter,
+            models.VegetationDrag.stem_count == None,
+            ~is_none_or_empty(models.VegetationDrag.stem_count_file),
+        ),
+        message="v2_vegetation_drag.stem_count is recommended as fallback value when using a stem_count_file.",
+    ),
+    RangeCheck(
+        error_code=507,
+        column=models.VegetationDrag.stem_diameter,
+        filters=vegetation_drag_filter,
+        min_value=0,
+        left_inclusive=False,
+    ),
+    QueryCheck(
+        error_code=508,
+        column=models.VegetationDrag.stem_diameter,
+        invalid=Query(models.VegetationDrag).filter(
+            vegetation_drag_filter,
+            models.VegetationDrag.stem_diameter == None,
+            is_none_or_empty(models.VegetationDrag.stem_diameter_file),
+        ),
+        message="v2_vegetation_drag.stem_diameter must be defined.",
+    ),
+    QueryCheck(
+        error_code=509,
+        level=CheckLevel.WARNING,
+        column=models.VegetationDrag.stem_diameter,
+        invalid=Query(models.VegetationDrag).filter(
+            vegetation_drag_filter,
+            models.VegetationDrag.stem_diameter == None,
+            ~is_none_or_empty(models.VegetationDrag.stem_diameter_file),
+        ),
+        message="v2_vegetation_drag.stem_diameter is recommended as fallback value when using a stem_diameter_file.",
+    ),
+    RangeCheck(
+        error_code=510,
+        column=models.VegetationDrag.drag_coefficient,
+        filters=vegetation_drag_filter,
+        min_value=0,
+        left_inclusive=False,
+    ),
+    QueryCheck(
+        error_code=511,
+        column=models.VegetationDrag.drag_coefficient,
+        invalid=Query(models.VegetationDrag).filter(
+            vegetation_drag_filter,
+            models.VegetationDrag.drag_coefficient == None,
+            is_none_or_empty(models.VegetationDrag.drag_coefficient_file),
+        ),
+        message="v2_vegetation_drag.drag_coefficient must be defined.",
+    ),
+    QueryCheck(
+        error_code=512,
+        level=CheckLevel.WARNING,
+        column=models.VegetationDrag.drag_coefficient,
+        invalid=Query(models.VegetationDrag).filter(
+            vegetation_drag_filter,
+            models.VegetationDrag.drag_coefficient == None,
+            ~is_none_or_empty(models.VegetationDrag.drag_coefficient_file),
+        ),
+        message="v2_vegetation_drag.drag_coefficient is recommended as fallback value when using a drag_coefficient_file.",
+    ),
+]
 
 ## 06xx: INFLOW
 for (surface, surface_map, filters) in [
@@ -1785,7 +1908,12 @@ RASTER_COLUMNS_FILTERS = [
         models.GlobalSetting.initial_groundwater_level_file,
         first_setting_filter & (models.GlobalSetting.groundwater_settings_id != None),
     ),
+    (models.VegetationDrag.height_file, vegetation_drag_filter),
+    (models.VegetationDrag.stem_count_file, vegetation_drag_filter),
+    (models.VegetationDrag.stem_diameter_file, vegetation_drag_filter),
+    (models.VegetationDrag.drag_coefficient_file, vegetation_drag_filter),
 ]
+
 CHECKS += [
     GDALAvailableCheck(
         error_code=700, level=CheckLevel.WARNING, column=models.GlobalSetting.dem_file
@@ -1943,6 +2071,31 @@ CHECKS += [
         error_code=798,
         column=models.GlobalSetting.dem_file,
         filters=first_setting_filter,
+    ),
+    ## 100xx: We continue raster checks from 10000
+    RasterRangeCheck(
+        error_code=10001,
+        column=models.VegetationDrag.height_file,
+        filters=vegetation_drag_filter,
+        min_value=0,
+    ),
+    RasterRangeCheck(
+        error_code=10002,
+        column=models.VegetationDrag.stem_count_file,
+        filters=vegetation_drag_filter,
+        min_value=0,
+    ),
+    RasterRangeCheck(
+        error_code=10003,
+        column=models.VegetationDrag.stem_diameter_file,
+        filters=vegetation_drag_filter,
+        min_value=0,
+    ),
+    RasterRangeCheck(
+        error_code=10004,
+        column=models.VegetationDrag.drag_coefficient_file,
+        filters=vegetation_drag_filter,
+        min_value=0,
     ),
 ]