+++ CHANGES.rst
@@ -11,6 +11,9 @@ Changelog of threedi-modelchecker
 - Added check 74: groundwater boundary types are not allowed on 1D boundary
   conditions.
 
+- Added groundwater 1D2D range checks for manholes, channels, and pipes for
+  exchange_thickness, hydraulic_conductivity_in, and hydraulic_conductivity_out.
+
 
 2.0.1 (2023-03-20)
 ------------------