+++ CHANGES.rst
@@ -10,6 +10,17 @@ Changelog of threedi-modelchecker
   error 1300 will be raised by the modelchecker. The allow-beta flag has been added
   to the CLI interface to disable this check temporarily.
 
+- Add errors and warnings for vegetation_drag input. Both rasters and global values.
+
+- Added check 73: groundwater boundaries are allowed only when there is
+  groundwater hydraulic conductivity.
+
+- Added check 74: groundwater boundary types are not allowed on 1D boundary
+  conditions.
+
+- Added groundwater 1D2D range checks for manholes, channels, and pipes for
+  exchange_thickness, hydraulic_conductivity_in, and hydraulic_conductivity_out.
+
 
 2.0.1 (2023-03-20)
 ------------------